requirements.txt
# --- Core API ---
fastapi==0.111.0
uvicorn[standard]==0.30.1
tiktoken

# --- LLM / AI ---
openai==1.35.7

# --- Environment Management ---
python-dotenv==1.0.1

# --- Validation / Models ---
pydantic==2.7.4

# --- Logging ---
loguru==0.7.2

# --- Async HTTP (future integrations / webhooks) ---
httpx==0.27.0

# --- Security & DNS ---
dnspython==2.6.1

# --- YAML Config ---
pyyaml==6.0.1

# --- Async File I/O ---
aiofiles==23.2.1

# --- Testing ---
pytest==8.2.2
pytest-asyncio==0.23.7

# --- Optional (recommended for dev quality) ---
black==24.4.2
isort==5.13.2
